+++ service-worker.js
@@ -14,6 +14,7 @@ const ASSETS = [
   './js/platform/shell/releaseNotesController.js',
   './js/platform/shell/feedbackController.js',
   './js/platform/shell/serviceWorkerController.js',
+  './js/platform/shell/performanceController.js',
   './js/core/centralStore.js',
   './js/core/domUpdate.js',
   './js/core/eventDelegation.js',